Replaced Line Search with Newton Raphson
The previous version of the ML algorithm was implemented with a line search algorithm which works well for estimating correlations but may become unstable for general Normal mean estimation (especially of the mean to be estimated is outside of the line search range). The current implementation uses a Newton Raphson approach which works well for all sigma/cutoff/x values.